pyproject.toml
[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "bycycle.webapi"
version = "0.7.dev0"
description = "byCycle web API"
license = "GPL-3.0+"
authors = ["Wyatt Baldwin <wyatt@bycycle.org>"]
homepage = "https://bycycle.org/"
repository = "https://github.com/bycycle-org/bycycle.webapi"

keywords = [
    "bicycle",
    "bike",
    "directions",
    "routing",
    "trip planner",
]

packages = [
    { include = "bycycle" },
]

include = [
    "AUTHORS",
    "CHANGELOG",
    "INSTALL",
    "LICENSE",
    "Makefile",
    "commands.py",
    "commands.yaml",
    "*.ini",
    "bycycle/webapi/commands.yaml",
    "bycycle/webapi/templates/**",
]

[tool.poetry.dependencies]
python = "^3.6"
"bycycle.core" = { path = "../bycycle.core/" }
cornice = "^4.0.1"
Mako = "^1.1.0"
plaster_pastedeploy = "^0.7"
pyramid = "^1.10.4"
pyramid_mako = "^1.1.0"
pyramid_retry = "^2.1"
pyramid_sqlalchemy = "^1.6"
pyramid_tm = "^2.3"
transaction = "^3.0.0"
"zope.sqlalchemy" = "^1.2"
runcommands = "^1.0a48"

[tool.poetry.dev-dependencies]
ansible = "*"
bpython = "*"
coverage = "*"
pyramid_debugtoolbar = "*"
waitress = "*"
WebTest = "*"

[tool.poetry.plugins."paste.app_factory"]
main = "bycycle.webapi:main"
